requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
cloudpickle==3.0.0
cohere==4.31
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
faiss-cpu==1.7.4
fastapi==0.104.0
fastavro==1.8.2
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.59.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iopath==0.1.10
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
jsonlines==4.0.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.5.8
networkx==3.2
nomic==2.0.13
numpy==1.26.1
omegaconf==2.3.0
onnxruntime==1.16.1
overrides==7.4.0
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
portalocker==2.8.2
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.2
submitit==1.5.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torchaudio==2.1.0
torchmetrics==0.10.3
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.8
websockets==12.0
win32-setctime==1.1.0
wonderwords==2.2.0
xformers==0.0.22.post4
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
